Added DynamoDB implementation based on SQLite
LocalDynamoDB relies on Amazon's DynamoDBLocal for the actual
implementation and only provides a wrapper that creates and handles a
temporary SQLite database.

Included is a repacked DynamoDBLocal.jar without all the dependencies
needed to run it as a web application as well as sqlite4java since
pulling it from maven messes with the native libraries.